type long removed

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 
 int _atoi(char *s)
 {
-	long number = 0;
+	int number = 0;
 	int i = 0, sign = 1;
 
     /*saut des espaces blancs et caractère alphabétique*/
